Added first learned file.
This should be the first general outline of how this will work. A
directory will correspond to an item in the list, which will be how each
item is organized further.

I have yet to add a human-consumable version of the list which links
each item to the relevant directory. I'm keeping the list just raw data
for now, as it should make it easier to expand it in raw format and have
that be used as the model for using it elsewhere (in a database, HTML
file, API, etc.).