Started work on VM in Python
I'm implementing a simple VM in Python. The specifications are from my
book in CS 160. The machine is very small and very simple. 16 registers,
256 memory cells, and 12 machine instructions. So far the basic
instructions are mostly functional. Jumping is not currently functional,
as I don't have a proper machine cycle implemented. Programs can be run
as follows:

    python machine.py 0 0x21AA 0x2255 0x9012 0xC000

The first argument (`0`) is the first address to use for storing the
program. Currently the only functional option is 0, as any other memory
cell will throw an error ('operator doesn't exist'). The following
arguments are machine instructions. The above program does the
following:

    LOAD R1 0xAA ;; (load register 1 with the value 0xAA)
    LOAD R2 0x55 ;; (load register 2 with the value 0x55)
    XOR R0 R1 R2 ;; (xor r1 and r2, store result in r0)
    HALT

Currently I have one "meta-instruction" implemented, `PRINT`. `PRINT` takes
options for which register and memory cell to print the values
of. `0xF5A3` will print the value of register 5 and the value of memory
cell 0xA3. This needs to be done before `HALT`ing the program, as `HALT`
is functional and truly ends the program execution when called. For
meaningful results in the above example, use the following invocation:

    python machine.py 0 0x21AA 0x2255 0x9012 0xF000 0xC000

Which will have a complete output of:

    R0: 0b11111111
    Cell 0x0: 0x21AA
    HALT