Implemented ROTATE
ROTATE will now correctly rotate a register's bits by up to 15 places to
the right, storing the bits at the front. Currently bits 4-8 are not
used in the operands, but perhaps I could use them as a true/false value
to keep the bits or not, although this would stray from the machine's
description in the textbook.