Started 05-Lists.idr Koans
Unfortunately it seems as though the Koans are becoming increasingly
outdated as they go on. I'm getting quite a number of warnings/errors(?)
for the skeletons provided and am having trouble figuring out how to
solve the errors. Perhaps a fork and a (hopeful) correction of these
errors would be a better exercise than the koans themselves.